% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetclean.R
\name{clean_df}
\alias{clean_df}
\title{Adds new column(s) to dataframe returned by rtweet get_timeline() function with
default parameters based on user specified input}
\usage{
clean_df(
  raw_tweets_df,
  text_only = TRUE,
  word_count = TRUE,
  emojis = TRUE,
  proportion_of_avg_retweets = TRUE,
  proportion_of_avg_favorites = TRUE
)
}
\arguments{
\item{raw_tweets_df}{dataframe}

\item{text_only}{bool}

\item{word_count}{bool}

\item{emojis}{bool}

\item{proportion_of_avg_retweets}{bool}

\item{proportion_of_avg_favorites}{bool}
}
\value{
dataframe
}
\description{
Returns a new dataframe containing additional columns that were not in the original
Generatable columns include...
text_only: strips emojis, hashtags, and hyperlinks from the text column
word_count: counts the number of words contained in the text_only column
emojis: contains the extracted emojis from text
proportion_of_avg_retweets: a proportion value of how many retweets a tweet received compared to the account average
proportion_of_avg_favorites: a proportion value of how many favorites a tweet received compared to the account average
}
\examples{
text <- c("example tweet text 1 @user2 @user", "#example #tweet 2 ", "example tweet 3 https://t.co/G4ziCaPond", "example tweet 4")
retweet_count <- c(43, 12, 24, 29)
favorite_count <- c(85, 41, 65, 54)
raw_df <- data.frame(text, retweet_count, favorite_count)
clean_df(raw_df)
clean_df(raw_df, emojis = FALSE)
}
