% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetclean.R
\name{engagement_by_hour}
\alias{engagement_by_hour}
\title{Average engagement by hour}
\usage{
engagement_by_hour(tweets_df)
}
\arguments{
\item{tweets_df}{dataframe}
}
\value{
ggplot object
}
\description{
Creates a line chart of average number of likes and retweets received based on hour of tweet posted.
}
