% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrc_emotion_lexicon.R
\docType{data}
\name{nrc_emotion_lexicon}
\alias{nrc_emotion_lexicon}
\title{Canadian-level language data collected from the 2016 Canadian Census.}
\format{
A tibble with 178 rows and 6 variables:
\describe{
\item{text_only}{Higher level language category, describing whether the language is an Official Canadian language (e.g., English or French), an Aboriginal language, or a Non-Official and Non-Aboriginal language.}
\item{sentiment}{Language queried about on the Canadian Census. Note: n.i.e. means Not included elsewhere, and n.o.s. means Not otherwise specified}
\item{count}{Total Count of Canadians from the Census who reported the language as their mother tongue (excluding institutional residents - 100\% data). Mother tongue is generally defined as the language someone was exposed to since birth.}
}
}
\source{
{Source: Statistics Canada, Census of Population, 2016. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.}
}
\usage{
nrc_emotion_lexicon
}
\description{
\code{nrc_emotion_lexicon} lists the counts of how many Canadians report each language as their mother tongue,
which language they speak most often at home, which language they use most often at work, and
which language they have knowledge for. This data was recorded in the 2016 Canadian Census.
}
\keyword{datasets}
