% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetclean.R
\name{sentiment_total}
\alias{sentiment_total}
\title{Sentiment Word Counts}
\usage{
sentiment_total(tweets, drop_sentiment = FALSE)
}
\arguments{
\item{tweets}{1-column dataframe}

\item{drop_sentiment}{A true/false bool that drops sentiment rows if no words are
associated with that sentiment}
}
\value{
dataframe
}
\description{
Takes an input of of single english words and outputs the number of words associated
with eight emotions and positive/negative sentiment. This is based on the the
crowd-sourced NRC Emotion Lexicon, which associates words with eight basic emotions
(anger, fear, anticipation, trust, surprise, sadness, joy, and disgust) and two
sentiments (negative and positive). For more information on NRC:
http://saifmohammad.com/WebPages/NRC-Emotion-Lexicon.htm
Note that words can be 0:n with emotions (either associated with none, 1, or many).
}
\examples{
tweets <- data.frame(word = c("this is example tweet 1",
                             "this is example tweet 2 with a few extra words",
                             "is third",
                             "4th tweet",
                             "fifth tweet"))
sentiment_total(tweets, drop_sentiment = FALSE)
}
