% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetclean.R
\name{tweet_words}
\alias{tweet_words}
\title{Most common words}
\usage{
tweet_words(clean_dataframe, top_n = 1)
}
\arguments{
\item{clean_dataframe}{dataframe}

\item{top_n}{numeric}
}
\value{
dataframe
}
\description{
Returns the top_n most common words and counts of occurrences from a list of tweets.
The output is sorted descending by the count of words and in reverse
alphabetical order for any word ties.
}
