% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweetclean.R
\name{tweet_words}
\alias{tweet_words}
\title{Most common words}
\usage{
tweet_words(clean_dataframe, top_n = 1)
}
\arguments{
\item{clean_dataframe}{dataframe}

\item{top_n}{numeric}
}
\value{
dataframe

data.frame
}
\description{
Returns the top_n most common words and counts of occurrences from a list of tweets.
The output is sorted descending by the count of words and in reverse
alphabetical order for any word ties.
}
\examples{
#' @param clean_dataframe data.frame
input_data <- (data.frame(id = c(1,2,3,4,5),
text_only = c("this is example tweet 1",
             "this is example tweet 2 with a few extra words",
             "is third",
             "4th tweet",
             "fifth tweet")))
tweet_words(input_data, 3)
}
